Extract an AddonClient for reuse
