+++ src/admin-base.ts
@@ -1,98 +1,44 @@
 import color from '@heroku-cli/color'
-import {APIClient, Command} from '@heroku-cli/command'
+import {Command} from '@heroku-cli/command'
 import cli from 'cli-ux'
-import * as _ from 'lodash'
-import * as url from 'url'
+
+import AddonClient from './addon-client'
 
 export default abstract class AdminBase extends Command {
   get addons() {
-    const client = new APIClient(this.config, {})
-    const host = process.env.HEROKU_ADDONS_HOST
-    client.defaults.host = host ? url.parse(host).host : 'addons.heroku.com'
-
-    let options = {
-      headers: {
-        'Content-Type': 'application/json',
-        Accept: 'application/json',
-        'User-Agent': 'kensa future'
-      }
-    }
+    const client = new AddonClient(this.config)
 
     const pull = async (slug: string) => {
-      try {
-        cli.action.start(`Fetching add-on manifest for ${color.addon(slug)}`)
-        const response = await client.get(`/api/v3/addons/${encodeURIComponent(slug)}/current_manifest`, options)
-        const body: any = response.body
-        cli.action.stop()
+      cli.action.start(`Fetching add-on manifest for ${color.addon(slug)}`)
+      const body = await client.get(`/api/v3/addons/${encodeURIComponent(slug)}/current_manifest`)
+      cli.action.stop()
 
-        return body.contents
-      } catch (err) {
-        const error = _.get(err, 'body.error')
-        if (error) {
-          this.error(error)
-        }
-        throw err
-      }
+      return body.contents
     }
 
     const manifests = async (slug: string) => {
-      try {
-        cli.action.start(`Fetching add-on manifests for ${color.addon(slug)}`)
-        const response = await client.get(`/api/v3/addons/${encodeURIComponent(slug)}/manifests`, options)
-        cli.action.stop()
+      cli.action.start(`Fetching add-on manifests for ${color.addon(slug)}`)
+      const body = await client.get(`/api/v3/addons/${encodeURIComponent(slug)}/manifests`)
+      cli.action.stop()
 
-        return response.body
-      } catch (err) {
-        const error = _.get(err, 'body.error')
-        if (error) {
-          this.error(error)
-        }
-        throw err
-      }
+      return body
     }
 
     const manifest = async (slug: string, uuid: string) => {
-      try {
-        cli.action.start(`Fetching add-on manifest for ${color.addon(slug)}`)
-        const response = await client.get(`/api/v3/addons/${encodeURIComponent(slug)}/manifests/${encodeURIComponent(uuid)}`, options)
-        const body: any = response.body
-        cli.action.stop()
+      cli.action.start(`Fetching add-on manifest for ${color.addon(slug)}`)
+      const body = await client.get(`/api/v3/addons/${encodeURIComponent(slug)}/manifests/${encodeURIComponent(uuid)}`)
+      cli.action.stop()
 
-        return body.contents
-      } catch (err) {
-        const error = _.get(err, 'body.error')
-        if (error) {
-          this.error(error)
-        }
-        throw err
-      }
+      return body.contents
     }
 
     const push = async (manifest: any) => {
       const requestBody = {contents: manifest}
-      let opts = {
-        ...options,
-        body: requestBody
-      }
+      cli.action.start('Pushing manifest')
+      const body = await client.post(`/api/v3/addons/${encodeURIComponent(manifest.id)}/manifests`, requestBody)
+      cli.action.stop()
 
-      try {
-        cli.action.start('Pushing manifest')
-        const response = await client.post(`/api/v3/addons/${encodeURIComponent(manifest.id)}/manifests`, opts)
-        const body: any = response.body
-        cli.action.stop()
-
-        return body.contents
-      } catch (err) {
-        const baseErrors = _.get(err, 'body.error.base')
-        if (baseErrors) {
-          this.error(baseErrors.join(', '))
-        }
-        const error = _.get(err, 'body.error')
-        if (error) {
-          this.error(error)
-        }
-        throw err
-      }
+      return body.contents
     }
 
     return {
